Update the names of updateStorage --> updateLocalStorage / getStorage --> getLocalStorage
