+++ src/main-src.js
@@ -3,14 +3,14 @@
 import './style-src.css';
 import Sortable from 'sortablejs';
 import {
-  showList, updateStorage, addTask, editTask, updateStatus,
+  showList, updateLocalStorage, addTask, editTask, updateStatus,
 } from './task.js';
 
 import {
-  getStorage, updateAfterDrag, removeCompletedTasks, removeTask,
+  getLocalStorage, updateAfterDrag, removeCompletedTasks, removeTask,
 } from './storage.js';
 
-let tasks = getStorage();
+let tasks = getLocalStorage();
 
 showList(tasks);
 
@@ -19,23 +19,23 @@ const clearAllButton = document.getElementById('clear-list');
 listDiv.addEventListener('click', (event) => {
   if (event.target !== event.currentTarget) {
     if (event.target.className === 'check') {
-      tasks = getStorage();
+      tasks = getLocalStorage();
       updateStatus(tasks, parseInt(event.target.parentElement.parentElement.id, 10), true);
       showList(tasks);
-      updateStorage(tasks);
+      updateLocalStorage(tasks);
     }
     if (event.target.className === 'fas fa-check') {
-      tasks = getStorage();
+      tasks = getLocalStorage();
       updateStatus(tasks, parseInt(event.target.parentElement.parentElement.id, 10), false);
       showList(tasks);
-      updateStorage(tasks);
+      updateLocalStorage(tasks);
     }
     if (event.target.className === 'fas fa-trash') {
-      tasks = getStorage();
+      tasks = getLocalStorage();
       tasks = removeTask(tasks, parseInt(event.target.parentElement.id, 10));
     }
     if (event.target.classList.contains('description')) {
-      tasks = getStorage();
+      tasks = getLocalStorage();
       editTask(tasks, event.target);
     }
   }
@@ -49,7 +49,7 @@ clearAllButton.addEventListener('click', () => {
 const addTaskButton = document.getElementById('add');
 addTaskButton.addEventListener('click', () => {
   addTask(tasks);
-  updateStorage(tasks);
+  updateLocalStorage(tasks);
   showList(tasks);
 });
